Add files via upload
This is my implementation of the original Viewie I coded. This time I made it nicer looking and included two optional viewing diagrams. In the future I would like to add labels for the Force Directed Graph right now its just nodes. As for the tree I havent put any complex code in to see the full diagram. I did not save it as a pdf mainly because I had a hard time. Currently still a png.